+++ docs/whats_new.rst
@@ -51,6 +51,7 @@ Enhancements
 - Ensure consistency in the last layer using tests (:gh:`642` by `Bruno Aristimunha`_ )
 - Ensuring consistency on the expose of the activation function (:gh:`637` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.SyncNet` (:gh:`643` by `Bruno Aristimunha`_ )
+- Adding :class:`braindecode.models.LDMANet` (:gh:`663` by `Bruno Aristimunha`_ )
 - Creating the `FilterBanklayer` module for new models (:gh:`656` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.IFNet` (:gh:`660` by `Bruno Aristimunha`_ )
 - Creating the FilterBank module for new models (:gh:`656` by `Bruno Aristimunha`_ )